Implements data corruption

+++ iLearning.PersonalDataRandomizer.Domain/Models/RandomOptions.cs
@@ -4,6 +4,6 @@ public class RandomOptions
 {
     public int Seed { get; set; }
     public string Country { get; set; }
-    public int ErrorsCount { get; set; }
+    public float ErrorsCount { get; set; }
     public int Size { get; set; } = 10;
 }
